Cleanup pretender setup, add coverage
