+++ package.json
@@ -40,7 +40,9 @@
     "html-webpack-plugin": "^3.2.0",
     "interactor.js": "^1.4.4",
     "karma": "^4.1.0",
+    "karma-coverage": "^1.1.2",
     "karma-firefox-launcher": "^1.1.0",
+    "karma-junit-reporter": "^1.2.0",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.5",
     "karma-webpack": "^4.0.2",
